requirements.txt
numpy==1.21.2
Pillow==8.4.0
torch==1.7.1+cu101
torchvision==0.8.2+cu101
